+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Version 0.1.0
 
 2019-05-03
 ----------
+- topmodel.py: Add pet calibration coefficient (KPEC) to parameters file
+  and update pet_hamon calculation to use the user-specified value.
+
+2019-05-03
+----------
 - topmodel.py: Add SCS runoff curve number calculation to hydrocals.py.
 
 2019-04-29